Handle clones always diffs against the original
Sometimes the passed-in data was an instance but had lost its reactivity for whatever reason.  This updates the handle-clones saveHandlers to always compare the clone against the original item in the store.  Without this change, it was impossible to set an item back to its original value, since the item passed in the props was non-reactive and always retained its original value instead of updating as it should.