fix($select) always return entire instance from store
There's not really a need for $select on the client. It is intended to decrease the amount of data returned from the server.  Plus, $select would result in non-instance partial copies of data as plain objects and not instances, so you'd lose the ability to call instance.save if we didn't return the instance.