fix disable $select on the client getter
The $select getter is causing an infinite loop when combined with the hydration logic.  We definitely want the getter to return full instances, so currently the hydration logic wins in order of importance.  Since $select is mostly useful for API fetches, I’m disabling it on the client to let it keep working with the server.